requirements.txt
streamlit
google-generativeai
tavily-python